[REFACTOR] 패키지명 rss 를 blog로 변경 (#1146)
* refactor: 패키지 이름 변경 (rss --> blog)

* chore: 로컬에서 otel 수집 기능 끄기 (otel export 에러 알림 발생 해결)
