+++ backend/baguni-batch/src/main/java/baguni/batch/domain/feed/service/FeedService.java
@@ -16,12 +16,12 @@ import baguni.batch.domain.feed.util.FeedApi;
 import baguni.batch.domain.feed.dto.AtomFeed;
 import baguni.batch.domain.feed.dto.Article;
 import baguni.batch.domain.feed.dto.RssFeed;
-import baguni.batch.infrastructure.feed.BlogDataHandler;
+import baguni.domain.infrastructure.blog.BlogDataHandler;
 import baguni.common.event.events.LinkCreateEvent;
 import baguni.common.event.messenger.EventMessenger;
 import baguni.domain.infrastructure.link.LinkDataHandler;
 import baguni.domain.model.link.Link;
-import baguni.domain.model.rss.Blog;
+import baguni.domain.model.blog.Blog;
 import io.opentelemetry.instrumentation.annotations.WithSpan;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
