+++ backend/baguni-batch/src/main/java/baguni/batch/listener/LinkUpdateEventListener.java
@@ -1,4 +1,4 @@
-package baguni.batch.domain.link.service;
+package baguni.batch.listener;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
@@ -8,6 +8,7 @@ import org.springframework.amqp.rabbit.annotation.RabbitHandler;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
 
+import baguni.batch.domain.link.service.LinkService;
 import baguni.common.config.RabbitmqConfig;
 import baguni.common.event.events.BookmarkCreateEvent;
 import baguni.common.event.events.LinkCreateEvent;
