+++ backend/baguni-domain/src/main/java/baguni/domain/infrastructure/blog/BlogDataHandler.java
@@ -1,4 +1,4 @@
-package baguni.batch.infrastructure.feed;
+package baguni.domain.infrastructure.blog;
 
 import java.util.List;
 
@@ -6,8 +6,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
-import baguni.domain.model.rss.Blog;
-import baguni.domain.infrastructure.rss.BlogRepository;
+import baguni.domain.model.blog.Blog;
 
 @Component
 @RequiredArgsConstructor
