+++ backend/baguni-domain/src/main/java/baguni/domain/infrastructure/blog/BlogRepository.java
@@ -1,8 +1,8 @@
-package baguni.domain.infrastructure.rss;
+package baguni.domain.infrastructure.blog;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import baguni.domain.model.rss.Blog;
+import baguni.domain.model.blog.Blog;
 
 public interface BlogRepository extends JpaRepository<Blog, Long> {
 }
